requirements.txt
Django==1.8.3
django-chartit==0.1
#django-debug-toolbar==1.4
django-object-tools==1.1.0
lxml==3.4.4
MySQL-python==1.2.5
Pillow==3.1.1
pingpp==2.0.7
pip==8.0.2
python-memcached==1.57
pytz==2015.7
requests==2.9.1
setuptools==20.1.1
simplejson==3.8.2
six==1.10.0
sqlparse==0.1.18
#suds==0.4
wheel==0.24.0
xlrd==1.0.0
xlwt==1.1.2
qiniu==7.0.6